refactor: move CSV escaping utility to src/utils
- Move escape_csv_field from scripts to src/utils for reusability
- Fix the broken CSV escaping implementation
- Add proper RFC 4180 compliant CSV field escaping
- Update ChemBL scraper import path
