+++ src/__init__.py
@@ -1,0 +1,3 @@
+"""
+Source code for the OSPF Ayurveda Knowledge Graph project.
+"""
